# api-docs-mon-archive.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/roles/{id}/permissions": {
      "put": {
        "tags": [
          "role-controller"
        ],
        "operationId": "updatePermissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "uniqueItems": true,
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{code}": {
      "get": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "findByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/roles": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "getAllRoles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "role-controller"
        ],
        "operationId": "createRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permissions": {
      "get": {
        "tags": [
          "permission-controller"
        ],
        "operationId": "getAllPermissions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "permission-controller"
        ],
        "operationId": "createPermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions": {
      "get": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MissionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{code}/equipe": {
      "get": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "getEquipeByMission",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "addMembreEquipe",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EquipeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{code}/assign": {
      "get": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "getAssignedUsers",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserMissionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "assignUser",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMissionAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserMissionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/roles/{roleName}/permissions": {
      "delete": {
        "tags": [
          "role-controller"
        ],
        "operationId": "removePermissions",
        "parameters": [
          {
            "name": "roleName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "uniqueItems": true,
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/roles/{roleName}/permissions/all": {
      "delete": {
        "tags": [
          "role-controller"
        ],
        "operationId": "removeAllPermissions",
        "parameters": [
          {
            "name": "roleName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/missions/{code}/assign/{userId}": {
      "delete": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "unassignUser",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
      "MissionUpdateRequest": {
        "type": "object",
        "properties": {
          "statut": {
            "type": "string",
            "enum": [
              "PLANIFIEE",
              "EN_COURS",
              "TERMINEE",
              "SUSPENDUE"
            ]
          },
          "exercice": {
            "type": "string"
          }
        }
      },
      "MissionResponse": {
        "type": "object",
        "properties": {
          "codeMission": {
            "type": "string"
          },
          "exercice": {
            "type": "string"
          },
          "codeDossier": {
            "type": "string"
          },
          "statut": {
            "type": "string",
            "enum": [
              "PLANIFIEE",
              "EN_COURS",
              "TERMINEE",
              "SUSPENDUE"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RoleRequest": {
        "required": [
          "name",
          "permissionNames"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "permissionNames": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PermissionRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "MissionCreateRequest": {
        "required": [
          "codeDossier",
          "codeMission",
          "exercice"
        ],
        "type": "object",
        "properties": {
          "codeMission": {
            "type": "string"
          },
          "exercice": {
            "type": "string"
          },
          "codeDossier": {
            "type": "string"
          }
        }
      },
      "EquipeCreateRequest": {
        "required": [
          "codeMission",
          "poste",
          "userId"
        ],
        "type": "object",
        "properties": {
          "codeMission": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "poste": {
            "type": "string",
            "enum": [
              "CHEF_MISSION",
              "SUPERVISEUR",
              "SENIOR",
              "JUNIOR",
              "ASSISTANT",
              "STAGIAIRE"
            ]
          }
        }
      },
      "EquipeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "codeMission": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "poste": {
            "type": "string",
            "enum": [
              "CHEF_MISSION",
              "SUPERVISEUR",
              "SENIOR",
              "JUNIOR",
              "ASSISTANT",
              "STAGIAIRE"
            ]
          }
        }
      },
      "UserMissionAssignRequest": {
        "required": [
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UserMissionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "missionId": {
            "type": "integer",
            "format": "int64"
          },
          "codeMission": {
            "type": "string"
          },
          "assignedBy": {
            "type": "integer",
            "format": "int64"
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}